Apply a patch to remove the built-in MIME database
We have a system-wide one, we don't need a copy in Qt, which is is at
best a copy of what's there, but needs decompressing, parsing and
memory allocation.

Needed to disable LTO since Clear is now enabling this for all
modules. Qtbase doesn't compile with -flto, it needs to control LTO on
its own.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>